input: use size_t for i_slave
i_slave represents an array size, so size_t is more suited than signed
integer for that.

+++ src/input/input_internal.h
@@ -522,7 +522,7 @@ typedef struct input_thread_private_t
     /* Main source */
     input_source_t *master;
     /* Slave sources (subs, and others) */
-    int            i_slave;
+    size_t         i_slave;
     input_source_t **slave;
     float          slave_subs_rate;
